trading_ai_system/requirements.txt
python-telegram-bot==20.3
crewai
langchain
langchain-openai
langchain-community
openai
chromadb
numpy<2.0
